Cache response digest header names
